Decouple the Actor impld type from the Requestor
The `Requestor` should be free from traces of the Actor implementation type
and shall be more concerned with the request and response types only.
This removes some complexity in generic implementations of an Actors
that has generic types.  The `Requestor` should be dissociated from it.

This should hopefully address an issue where the Requestor is not easily
cloned due to it needing to carry the Actor type to get its request and
response types.