Bump to v0.2
This is a breaking change from 0.1.x versions in the signature change in
the `Requestor` struct.  This is a necessary change to simplify the
client code as it removes the `Requestor` from carrying any complex
generic types that the impld `Actor` may carry.

The `Requestor` should only be concerned about the request and response
types that is connecting the client and the actor itself.